Rename eventloop.h to EventLoop.h
We usually use upper camel casing for filenames.
